packages/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
httpx==0.27.2
boto3==1.35.0
python-dotenv==1.0.1
claude-agent-sdk
pydantic==2.8.2
pytest==8.3.3
pytest-asyncio==0.24.0
